refactor(encoding): Phase 1e Stage C batch 17 — lift optimal-parser pure helpers onto impl BtMatcher
Move all 11 stateless optimal-parser helpers off `impl HcMatchGenerator`
onto `impl BtMatcher`:

- `encode_offset_with_reps`, `encode_offset_base_with_reps`
- `update_plan_stats_segment`
- `reset_opt_node`, `reset_opt_nodes`
- `add_price_delta`, `add_prices`
- `cached_literal_price`, `cached_lit_length_price`,
  `cached_lit_length_delta_price`, `cached_match_length_price`

All are pure associated functions (no `self`), so the migration is a
relocation only — internal `Self::` recursion (e.g.
`cached_lit_length_delta_price` calling `cached_lit_length_price`)
keeps working unchanged.

~37 caller sites across `build_optimal_plan_impl!` and
`run_btultra2_seed_pass` / `start_matching_optimal` rewritten from
`HcMatchGenerator::X` / `Self::X` to `BtMatcher::X`.

`BtMatcher` added to imports; `HC_MAX_LIT` gated to `#[cfg(test)]`
since the production use moved to `bt/mod.rs`.

Build (default + --no-default-features), strict clippy, 381 lib tests
incl. level22 ratio gate — all green.
